Indent after a change or delete, so that indenation is preserved.
If the block is unbalanced, the dd operation is a no-op, but the
mechanics of evil-delete sets the starting point of the motion
(the whole line) as the beginning of the line, and thus the unbalanced sexps
get deindented.

Another fix would be to modify evil-sp--new-beginning to use the start of the
current sexp as the beginning of the motion. This is probably a bit smarter,
as the current feature feels a bit hairy, though it works.