[Fix #20] Don't delete space after word with cw
This fixes a discrepency with how evil and evil-smartparens works.
evil-sp would delete the space after words when issuing a cw command,
which is clearly suboptimal because you'd then have to re-enter that space
after changing the word.

The solution is a bit hackish, but I don't want to start digging into the
internals of evil again just to solve this relatively minor issue.