Keep bundled evil-tests.el of the load-path
Bundling a third-party library should generally be avoided because the
bundled version might end up being loaded for users who also have a
more recent version installed separately.

But because a user who wants to run the tests might not have `evil`'s
test file handy, it might make sense to bundle a copy anyway.  This
commit puts `evil-tests.el` in a separate directory which also
contains a file `.nosearch`, which ensures that it is not added to the
`load-path`.  The bundled library is still loaded, but only if the
upstream library cannot be found otherwise.